Fixed issue with matching signature columns
The Exact and force matching_signature columns of v$sql are of NUMBER datatype, but they really are long numeric strings. Previous versions were fetching them into C int vars. This problem appeared with newest versions of Oracle (18c). The fields in response_times were modified to VARCHAR2(40), along with the C variables, so as to not have an issue anymore